requirements.txt
torch
torchvision
scikit-learn
pandas
opencv-python
Pillow
dlib
